webapi/tct-screenorientation-w3c-tests/screenorientation/Screen_lockOrientation_parameter_not_in_allowed_orientations.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xinx, liu <xinx.liu@intel.com>
        Peng,keli <kelix.peng@intel.com>

-->
<html>
  <head>
    <title>Screen Test:Screen_lockOrientation_parameter_not_in_allowed_orientations</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-screen-orientation-20120522/#extensions-to-the-screen-interface" />
    <meta name="flags" content="interact" />
    <meta name="assert" content="Check  the method lockOrientation  of interface Screen when parameter is not in allowed orientations" />
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script type="text/javascript">
        var t = async_test(document.title);
        var allowed_orientations = new Array("portrait","landscape","portrait-primary","portrait-secondary","landscape-primary","landscape-secondary")
        function show() {
            try {
                var orientation="dsdsf";
                //parameter orientation  not in allowed_orientations
                var aa=screen.lockOrientation(orientation);
                t.step(function () {
                    assert_true(aa == false,"Check  the method lockOrientation  of interface Screen when parameter is not in allowed orientations");
                });
                t.done();
            } catch (err) {
                t.step(function () {
                    assert_true(err.name == "NOT_SUPPORTED_ERROR","Mark the case as \"PASS\" because the platform actually doesn't support the interface");
                });
                t.done();
            }
        }
        show();
    </script>
    <div id="log"></div>
  </body>
</html>
